requirements.txt
xlrd==1.2.0
pyyaml
openai==0.28
tqdm
numpy
numba
evaluate>=0.3.0
absl-py
nltk
rouge-score
pandas
torch
opencv-python
matplotlib
Pillow
plotly
kaleido
rpy2
requests
openpyxl
yolk
xlsxwriter